長期間空いたので確認とgameObject親クラスを作成 それに準拠するように変更

+++ RPG_3D_Game/GameInfo.h
@@ -1,3 +1,3 @@
 #pragma once
-const double WIN_SIZE_X = 1280.0;	// 画面サイズX
-const double WIN_SIZE_Y = 720.0;	// 画面サイズY
+const float WIN_SIZE_X = 1280.0;	// 画面サイズX
+const float WIN_SIZE_Y = 720.0;	// 画面サイズY
